feat: wire CalendarPanel into main app and highlight due dates
- Move CalendarPanel.java into src/ so Ant compiles it.
- Accept DefaultListModel<Subject> and read the existing "Due Date"
  column instead of introducing a parallel Assignment model.
- Mark cells with due items with a red dot, count, and tooltip.
- Add a JTabbedPane (Assignments / Calendar) to the main frame and
  refresh the calendar on tab switch and after add/remove/rename/delete.

Closes #38
